misc/HtmlTestSuite/results/test-0912.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>document.write</title>
<script src="/resources/testharness.js"/><script src="/resources/testharnessreport.js"/>
<script>
var t = async_test();
t.step(function() {
  var s = "&lt;body&gt;&lt;textarea&gt;&lt;span&gt;Filler&lt;/span&gt;&lt;/textarea&gt;";
  for (var i=0; i&lt;s.length; i++) {
    document.write(s[i]);
  }
});
</script>
<script>
t.step(function() {
  assert_equals(document.body.childNodes[0].localName, "textarea");
  assert_equals(document.body.childNodes[0].textContent, "&lt;span&gt;Filler&lt;/span&gt;");
})
t.done();
</script>
</head><body><div id="log"/>
</body></html>
